pyproject.toml
[project]
name = "ideal-bot"
version = "0.1.0"
description = "Discord bot that speaks using registered vocabulary"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.3",
    "aiosqlite>=0.19",
    "pykakasi>=2.2",
    "cryptography>=42",
    "openai>=1.0",
    "google-genai>=1.0",
    "sentence-transformers>=3.0",
    "sentencepiece>=0.1.99",
    "transformers>=5.6",
    "accelerate>=1.0",
    "kernels>=0.12.0",
    "tiktoken>=0.12.0",
    "protobuf>=7.34.1",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "itsdangerous>=2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
